+++ psRadiantOne/Public/JdbcDrivers/Get-R1JdbcDriverLibrary.ps1
@@ -1,0 +1,36 @@
+# .ExternalHelp psRadiantOne-help.xml
+function Get-R1JdbcDriverLibrary {
+	[CmdletBinding()]
+	[OutputType('psRadiantOne.LibraryReference')]
+	param(
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[string]$name
+	)
+
+	Begin {
+
+		Assert-R1Session -RequireToken
+
+	}#begin
+
+	Process {
+
+		$URI = Resolve-R1ServiceUrl -Service Catalog -Path "jdbc_drivers/files/$($name | Get-EscapedString)/libraries"
+
+		$Result = Invoke-R1RestMethod -Uri $URI -Method GET
+
+		if ($null -ne $Result) {
+
+			$Result | Add-CustomType -Type psRadiantOne.LibraryReference
+
+		}
+
+	}#process
+
+	End { }#end
+
+}